+++ examples/basic/src/config.ts
@@ -1,4 +1,5 @@
 import { ConfigBuilder } from "@layerfig/config";
+import { FileSource } from "@layerfig/config/sources/file";
 
 export const config = new ConfigBuilder({
 	validate: (finalConfig, z) => {
@@ -9,6 +10,6 @@ export const config = new ConfigBuilder({
 		return schema.parse(finalConfig);
 	},
 })
-	.addSource("base.json")
-	.addSource("prod.json")
+	.addSource(new FileSource("base.json"))
+	.addSource(new FileSource("prod.json"))
 	.build();