+++ examples/json5/src/config.ts
@@ -1,4 +1,5 @@
 import { ConfigBuilder } from "@layerfig/config";
+import { FileSource } from "@layerfig/config/sources/file";
 import json5Parser from "@layerfig/parser-json5";
 
 export const config = new ConfigBuilder({
@@ -11,7 +12,7 @@ export const config = new ConfigBuilder({
 	},
 	parser: json5Parser,
 })
-	.addSource("base.jsonc")
-	.addSource("local.json")
-	.addSource("prod.json5")
+	.addSource(new FileSource("base.jsonc"))
+	.addSource(new FileSource("local.json"))
+	.addSource(new FileSource("prod.json5"))
 	.build();