+++ examples/toml/src/config.ts
@@ -1,4 +1,5 @@
 import { ConfigBuilder } from "@layerfig/config";
+import { FileSource } from "@layerfig/config/sources/file";
 import tomlParser from "@layerfig/parser-toml";
 
 export const config = new ConfigBuilder({
@@ -11,7 +12,7 @@ export const config = new ConfigBuilder({
 	},
 	parser: tomlParser,
 })
-	.addSource("base.toml")
-	.addSource("local.toml")
-	.addSource("prod.toml")
+	.addSource(new FileSource("base.toml"))
+	.addSource(new FileSource("local.toml"))
+	.addSource(new FileSource("prod.toml"))
 	.build();