+++ packages/config/tsdown.config.ts
@@ -1,7 +1,12 @@
 import { defineConfig } from "tsdown";
 
 export default defineConfig({
-	entry: ["./src/index.ts"],
+	entry: {
+		index: "./src/index.ts",
+		"sources/env/index": "./src/sources/env-var.ts",
+		"sources/file/index": "./src/sources/file.ts",
+		"sources/object/index": "./src/sources/object.ts",
+	},
 	dts: true,
 	sourcemap: true,
 	format: ["cjs", "esm"],