+++ packages/parser-yaml/src/index.test.ts
@@ -1,27 +1,12 @@
-import { ConfigBuilder } from "@layerfig/config";
+import { ConfigBuilder, type ConfigBuilderOptions } from "@layerfig/config";
+import { FileSource } from "@layerfig/config/sources/file";
 import { describe, expect, it } from "vitest";
 import { z } from "zod/v4";
 import yamlParser from "./index";
 
-const schema = z.object({
-	appURL: z.url(),
-	api: z.object({
-		port: z
-			.string()
-			.transform((val) => Number.parseInt(val, 10))
-			.or(z.number()),
-	}),
-});
-
 describe("yamlParser", () => {
-	const config = new ConfigBuilder({
-		validate: (config) => schema.parse(config),
-		configFolder: "./src/__fixtures__",
-		parser: yamlParser,
-	});
-
 	it("should load config from .yaml file", () => {
-		const result = config.addSource("base.yaml").build();
+		const result = getConfig().addSource(new FileSource("base.yaml")).build();
 
 		expect(result).toEqual({
 			appURL: "https://my-site.com",
@@ -32,7 +17,7 @@ describe("yamlParser", () => {
 	});
 
 	it("should load config from .yml file", () => {
-		const result = config.addSource("base.yml").build();
+		const result = getConfig().addSource(new FileSource("base.yml")).build();
 
 		expect(result).toEqual({
 			appURL: "https://my-site.com",
@@ -43,26 +28,53 @@ describe("yamlParser", () => {
 	});
 
 	it("should add layer json files", () => {
-		expect(config.addSource("base.yaml").addSource("dev.yaml").build()).toEqual(
-			{
-				appURL: "https://dev.company-app.com",
-				api: {
-					port: 3000,
-				},
+		expect(
+			getConfig()
+				.addSource(new FileSource("base.yaml"))
+				.addSource(new FileSource("dev.yaml"))
+				.build(),
+		).toEqual({
+			appURL: "https://dev.company-app.com",
+			api: {
+				port: 3000,
 			},
-		);
+		});
 
-		expect(config.addSource("dev.yaml").addSource("base.yaml").build()).toEqual(
-			{
-				appURL: "https://my-site.com",
-				api: {
-					port: 3000,
-				},
+		expect(
+			getConfig()
+				.addSource(new FileSource("dev.yaml"))
+				.addSource(new FileSource("base.yaml"))
+				.build(),
+		).toEqual({
+			appURL: "https://my-site.com",
+			api: {
+				port: 3000,
 			},
-		);
+		});
 	});
 
 	it("should throw an error if the file extension is not supported", () => {
-		expect(() => config.addSource("base.json").build()).toThrowError();
+		expect(() =>
+			getConfig().addSource(new FileSource("base.json")).build(),
+		).toThrowError();
 	});
 });
+
+function getConfig(options?: Partial<ConfigBuilderOptions>) {
+	const schema = z.object({
+		appURL: z.url(),
+		api: z.object({
+			port: z
+				.string()
+				.transform((val) => Number.parseInt(val, 10))
+				.or(z.number()),
+		}),
+	});
+
+	return new ConfigBuilder({
+		validate: (config) => schema.parse(config),
+		configFolder: "./src/__fixtures__",
+		parser: yamlParser,
+		...options,
+	});
+}